backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic[email]
email-validator
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
